requirements.txt
apply==1.6
certifi==2022.12.7
charset-normalizer==3.0.1
Chiplotle3==0.4.2
freetype-py==2.3.0
future==0.18.3
idna==3.4
numpy==1.24.2
oauthlib==3.2.2
Pillow==9.4.0
pyserial==3.5
requests==2.28.2
requests-oauthlib==1.3.1
scipy==1.10.0
svgpathtools==1.5.1
svgwrite==1.4.3
tweepy==4.12.1
urllib3==1.26.14
Wand==0.6.11
